Moved addr_lib to libcfnet
