+++ libcfnet/Makefile.am
@@ -29,15 +29,16 @@ AM_CPPFLAGS += -I$(top_srcdir)/libutils			    # platform.h
 AM_CPPFLAGS += -I$(top_srcdir)/libpromises		    # cf3.extern.h
 
 libcfnet_la_SOURCES = \
-	misc.c \
-	net.c net.h \
-	communication.c communication.h \
+	addr_lib.c addr_lib.h \
 	client_protocol.c client_protocol.h cfnet.h\
 	client_code.c client_code.h \
 	classic.c classic.h \
-	tls_client.c tls_client.h \
-	tls_generic.c tls_generic.h \
+	communication.c communication.h \
 	connection_info.c connection_info.h \
 	conn_cache.c conn_cache.h \
+	key.c key.h \
+	misc.c \
+	net.c net.h \
 	stat_cache.c stat_cache.h \
-	key.c key.h
+	tls_client.c tls_client.h \
+	tls_generic.c tls_generic.h
